chore(test-data): use one typefile per model
