+++ models/product-discount-value-absolute/src/product-discount-value-absolute-draft/builder.ts
@@ -4,7 +4,7 @@ import transformers from './transformers';
 import type {
   TCreateProductDiscountValueAbsoluteDraftBuilder,
   TProductDiscountValueAbsoluteDraft,
-} from './types';
+} from '../types';
 
 const Model: TCreateProductDiscountValueAbsoluteDraftBuilder = () =>
   Builder<TProductDiscountValueAbsoluteDraft>({
