+++ models/product-discount-value-absolute/src/product-discount-value-absolute-draft/generator.ts
@@ -1,5 +1,5 @@
 import { fake, Generator } from '@commercetools-test-data/core';
-import { TProductDiscountValueAbsoluteDraft } from './types';
+import { TProductDiscountValueAbsoluteDraft } from '../types';
 import * as Money from '@commercetools-test-data/money';
 
 // https://docs.commercetools.com/api/projects/productDiscounts#productdiscountvalueabsolutedraft
