+++ models/product-discount-value-absolute/src/product-discount-value-absolute-draft/transformers.ts
@@ -2,7 +2,7 @@ import { Transformer } from '@commercetools-test-data/core';
 import type {
   TProductDiscountValueAbsoluteDraft,
   TProductDiscountValueAbsoluteDraftGraphql,
-} from './types';
+} from '../types';
 
 const transformers = {
   default: Transformer<
