--- models/product-discount-value-absolute/src/product-discount-value-absolute-draft/types.ts
@@ -1,15 +1,0 @@
-import type { TBuilder } from '@commercetools-test-data/core';
-import { ProductDiscountValueAbsoluteDraft } from '@commercetools/platform-sdk';
-
-export type TProductDiscountValueAbsoluteDraft =
-  ProductDiscountValueAbsoluteDraft;
-
-export type TProductDiscountValueAbsoluteDraftGraphql =
-  TProductDiscountValueAbsoluteDraft & {
-    __typename: 'ProductDiscountValueAbsoluteDraft';
-  };
-
-export type TProductDiscountValueAbsoluteDraftBuilder =
-  TBuilder<TProductDiscountValueAbsoluteDraft>;
-export type TCreateProductDiscountValueAbsoluteDraftBuilder =
-  () => TProductDiscountValueAbsoluteDraftBuilder;
