+++ models/product-discount-value-external/src/product-discount-value-external-draft/builder.spec.ts
@@ -4,7 +4,7 @@ import { createBuilderSpec } from '@commercetools-test-data/core/test-utils';
 import {
   TProductDiscountValueExternalDraft,
   TProductDiscountValueExternalDraftGraphql,
-} from './types';
+} from '../types';
 import * as ProductDiscountValueExternalDraft from '.';
 
 describe('builder', () => {
