--- models/product-discount-value-external/src/product-discount-value-external-draft/types.ts
@@ -1,15 +1,0 @@
-import type { TBuilder } from '@commercetools-test-data/core';
-import { ProductDiscountValueExternalDraft } from '@commercetools/platform-sdk';
-
-export type TProductDiscountValueExternalDraft =
-  ProductDiscountValueExternalDraft;
-
-export type TProductDiscountValueExternalDraftGraphql =
-  TProductDiscountValueExternalDraft & {
-    __typename: 'ProductDiscountValueExternalDraft';
-  };
-
-export type TProductDiscountValueExternalDraftBuilder =
-  TBuilder<TProductDiscountValueExternalDraft>;
-export type TCreateProductDiscountValueExternalDraftBuilder =
-  () => TProductDiscountValueExternalDraftBuilder;
