+++ models/product-discount-value-relative/src/product-discount-value-relative-draft/builder.spec.ts
@@ -4,7 +4,7 @@ import { createBuilderSpec } from '@commercetools-test-data/core/test-utils';
 import {
   TProductDiscountValueRelativeDraft,
   TProductDiscountValueRelativeDraftGraphql,
-} from './types';
+} from '../types';
 import * as ProductDiscountValueRelativeDraft from '.';
 
 describe('builder', () => {
