+++ models/product-discount-value-relative/src/product-discount-value-relative-draft/builder.ts
@@ -4,7 +4,7 @@ import transformers from './transformers';
 import type {
   TCreateProductDiscountValueRelativeDraftBuilder,
   TProductDiscountValueRelativeDraft,
-} from './types';
+} from '../types';
 
 const Model: TCreateProductDiscountValueRelativeDraftBuilder = () =>
   Builder<TProductDiscountValueRelativeDraft>({
