+++ models/product-discount-value-relative/src/product-discount-value-relative-draft/transformers.ts
@@ -2,7 +2,7 @@ import { Transformer } from '@commercetools-test-data/core';
 import type {
   TProductDiscountValueRelativeDraft,
   TProductDiscountValueRelativeDraftGraphql,
-} from './types';
+} from '../types';
 
 const transformers = {
   default: Transformer<
