--- models/product-discount-value-relative/src/product-discount-value-relative-draft/types.ts
@@ -1,15 +1,0 @@
-import type { TBuilder } from '@commercetools-test-data/core';
-import { ProductDiscountValueRelativeDraft } from '@commercetools/platform-sdk';
-
-export type TProductDiscountValueRelativeDraft =
-  ProductDiscountValueRelativeDraft;
-
-export type TProductDiscountValueRelativeDraftGraphql =
-  TProductDiscountValueRelativeDraft & {
-    __typename: 'ProductDiscountValueRelativeDraft';
-  };
-
-export type TProductDiscountValueRelativeDraftBuilder =
-  TBuilder<TProductDiscountValueRelativeDraft>;
-export type TCreateProductDiscountValueRelativeDraftBuilder =
-  () => TProductDiscountValueRelativeDraftBuilder;
